add ListViewFocusController
